tr2/lara/flare: move flare control to Lara scope
This moves Flare_DoInHand to Lara's scope, along with its
dependencies. This will later be made private and renamed after
further refactors in gun.c.
