+++ src/tr2/decomp/flares.h
@@ -22,6 +22,5 @@ typedef enum {
 void Flare_GenerateEffects(
     XYZ_32 sound_pos, XYZ_32 flare_pos, int16_t room_num);
 bool Flare_GenerateLight(XYZ_32 pos, int32_t flare_age);
-void Flare_DoInHand(int32_t flare_age);
 void Flare_Create(bool thrown);
 int32_t Flare_GetMaxAge(void);
