+++ src/tr2/game/lara/flare.h
@@ -2,6 +2,7 @@
 
 #include <stdint.h>
 
+void Lara_Flare_DoInHand(int32_t flare_age);
 void Lara_Flare_SetArm(int32_t flare_frame);
 void Lara_Flare_DrawMeshes(void);
 void Lara_Flare_UndrawMeshes(void);
